--- solutions/standard/catalogValidationValues.json.template
@@ -1,8 +1,0 @@
-{
-  "ibmcloud_api_key": $VALIDATION_APIKEY,
-  "region": "us-south",
-  "tags": $TAGS,
-  "name": $PREFIX,
-  "resource_group_name": $PREFIX,
-  "existing_kms_instance_crn": $HPCS_US_SOUTH_CRN
-}
