--- solutions/standard/version.tf
@@ -1,18 +1,0 @@
-terraform {
-  required_version = ">= 1.3.0"
-  # Lock DA into an exact provider version - renovate automation will keep it updated
-  required_providers {
-    ibm = {
-      source  = "IBM-Cloud/ibm"
-      version = "1.74.0"
-    }
-    time = {
-      source  = "hashicorp/time"
-      version = "0.12.1"
-    }
-    random = {
-      source  = "hashicorp/random"
-      version = "3.6.3"
-    }
-  }
-}
