+++ tests/pr_test.go
@@ -4,24 +4,16 @@ package test
 import (
 	"crypto/rand"
 	"encoding/base64"
-	"encoding/json"
-	"fmt"
-	"io/fs"
 	"log"
 	"os"
-	"path/filepath"
-	"strings"
 	"testing"
 
+	"github.com/gruntwork-io/terratest/modules/terraform"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/cloudinfo"
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/common"
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/testhelper"
-	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/testschematic"
 )
-
-const standardSolutionTerraformDir = "solutions/standard"
 
 // Use existing resource group
 const resourceGroup = "geretain-test-enterprisedb"
@@ -49,153 +41,74 @@ func TestMain(m *testing.M) {
 	os.Exit(m.Run())
 }
 
-type tarIncludePatterns struct {
-	excludeDirs []string
-
-	includeFiletypes []string
-
-	includeDirs []string
-}
-
-func getTarIncludePatternsRecursively(dir string, dirsToExclude []string, fileTypesToInclude []string) ([]string, error) {
-	r := tarIncludePatterns{dirsToExclude, fileTypesToInclude, nil}
-	err := filepath.WalkDir(dir, func(path string, entry fs.DirEntry, err error) error {
-		return walk(&r, path, entry, err)
-	})
-	if err != nil {
-		fmt.Println("error")
-		return r.includeDirs, err
-	}
-	return r.includeDirs, nil
-}
-
-func walk(r *tarIncludePatterns, s string, d fs.DirEntry, err error) error {
-	if err != nil {
-		return err
-	}
-	if d.IsDir() {
-		for _, excludeDir := range r.excludeDirs {
-			if strings.Contains(s, excludeDir) {
-				return nil
-			}
-		}
-		if s == ".." {
-			r.includeDirs = append(r.includeDirs, "*.tf")
-			return nil
-		}
-		for _, includeFiletype := range r.includeFiletypes {
-			r.includeDirs = append(r.includeDirs, strings.ReplaceAll(s+"/*"+includeFiletype, "../", ""))
-		}
-	}
-	return nil
-}
-
-func TestRunStandardSolutionSchematics(t *testing.T) {
+func TestRunFSCloudExample(t *testing.T) {
 	t.Parallel()
 
-	excludeDirs := []string{
-		".terraform",
-		".docs",
-		".github",
-		".git",
-		".idea",
-		"common-dev-assets",
-		"examples",
-		"tests",
-		"reference-architectures",
-	}
-	includeFiletypes := []string{
-		".tf",
-		".yaml",
-		".py",
-		".tpl",
-		".sh",
-	}
-
-	tarIncludePatterns, recurseErr := getTarIncludePatternsRecursively("..", excludeDirs, includeFiletypes)
-
-	// if error producing tar patterns (very unexpected) fail test immediately
-	require.NoError(t, recurseErr, "Schematic Test had unexpected error traversing directory tree")
-	prefix := "edb-st-da"
-	options := testschematic.TestSchematicOptionsDefault(&testschematic.TestSchematicOptions{
-		Testing:                t,
-		TarIncludePatterns:     tarIncludePatterns,
-		TemplateFolder:         standardSolutionTerraformDir,
-		BestRegionYAMLPath:     regionSelectionPath,
-		Prefix:                 prefix,
-		ResourceGroup:          resourceGroup,
-		DeleteWorkspaceOnFail:  false,
-		WaitJobCompleteMinutes: 60,
+	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
+		Testing:      t,
+		TerraformDir: "examples/fscloud",
+		Prefix:       "edb-fscloud",
+		Region:       "us-south", // For FSCloud locking into us-south since that is where the HPCS permanent instance is
+		/*
+		 Comment out the 'ResourceGroup' input to force this tests to create a unique resource group to ensure tests do
+		 not clash. This is due to the fact that an auth policy may already exist in this resource group since we are
+		 re-using a permanent HPCS instance. By using a new resource group, the auth policy will not already exist
+		 since this module scopes auth policies by resource group.
+		*/
+		//ResourceGroup: resourceGroup,
+		TerraformVars: map[string]interface{}{
+			"access_tags":                permanentResources["accessTags"],
+			"existing_kms_instance_guid": permanentResources["hpcs_south"],
+			"kms_key_crn":                permanentResources["hpcs_south_root_key_crn"],
+			"edb_version":                "12", // Always lock this test into the latest supported Enterprise DB version
+		},
+		CloudInfoService: sharedInfoSvc,
 	})
 
-	serviceCredentialSecrets := []map[string]interface{}{
-		{
-			"secret_group_name": fmt.Sprintf("%s-secret-group", prefix),
-			"service_credentials": []map[string]string{
-				{
-					"secret_name": fmt.Sprintf("%s-cred-reader", prefix),
-					"service_credentials_source_service_role": "Reader",
-				},
-				{
-					"secret_name": fmt.Sprintf("%s-cred-writer", prefix),
-					"service_credentials_source_service_role": "Writer",
-				},
-			},
-		},
-	}
-
-	serviceCredentialNames := map[string]string{
-		"admin": "Administrator",
-		"user1": "Viewer",
-		"user2": "Editor",
-	}
-
-	serviceCredentialNamesJSON, err := json.Marshal(serviceCredentialNames)
-	if err != nil {
-		log.Fatalf("Error converting to JSON: %s", err)
-	}
-
-	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
-		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
-		{Name: "access_tags", Value: permanentResources["accessTags"], DataType: "list(string)"},
-		{Name: "existing_kms_instance_crn", Value: permanentResources["hpcs_south_crn"], DataType: "string"},
-		{Name: "kms_endpoint_type", Value: "private", DataType: "string"},
-		{Name: "edb_version", Value: "12", DataType: "string"}, // Always lock this test into the latest supported EnterpriseDB version
-		{Name: "resource_group_name", Value: options.Prefix, DataType: "string"},
-		{Name: "existing_secrets_manager_instance_crn", Value: permanentResources["secretsManagerCRN"], DataType: "string"},
-		{Name: "service_credential_secrets", Value: serviceCredentialSecrets, DataType: "list(object)"},
-		{Name: "service_credential_names", Value: string(serviceCredentialNamesJSON), DataType: "map(string)"},
-	}
-	err = options.RunSchematicTest()
+	options.SkipTestTearDown = true
+	output, err := options.RunTestConsistency()
 	assert.Nil(t, err, "This should not have errored")
+	assert.NotNil(t, output, "Expected some output")
+
+	// check for outputs
+	outputs := terraform.OutputAll(options.Testing, options.TerraformOptions)
+	expectedOutputs := []string{"port", "hostname"}
+	_, outputErr := testhelper.ValidateTerraformOutputs(outputs, expectedOutputs...)
+	assert.NoErrorf(t, outputErr, "Some outputs not found or nil")
+	options.TestTearDown()
 }
 
-func TestRunStandardUpgradeSolution(t *testing.T) {
+func TestRunUpgradeCompleteExample(t *testing.T) {
 	t.Parallel()
 
 	// Generate a 15 char long random string for the admin_pass
 	randomBytes := make([]byte, 13)
-	_, randErr := rand.Read(randomBytes)
-	require.Nil(t, randErr) // do not proceed if we can't gen a random password
+	_, err := rand.Read(randomBytes)
+	if err != nil {
+		log.Fatal(err)
+	}
 
 	randomPass := "A1" + base64.URLEncoding.EncodeToString(randomBytes)[:13]
 
-	options := testhelper.TestOptionsDefault(&testhelper.TestOptions{
+	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
 		Testing:            t,
-		TerraformDir:       standardSolutionTerraformDir,
+		TerraformDir:       "examples/complete",
+		Prefix:             "edb-upg",
 		BestRegionYAMLPath: regionSelectionPath,
-		Prefix:             "edb-st-da-upg",
 		ResourceGroup:      resourceGroup,
+		TerraformVars: map[string]interface{}{
+			"edb_version": "12", // Always lock to the lowest supported EDB version
+			"users": []map[string]interface{}{
+				{
+					"name":     "testuser",
+					"password": randomPass, // pragma: allowlist secret
+					"type":     "database",
+				},
+			},
+			"admin_pass": randomPass, // pragma: allowlist secret
+		},
+		CloudInfoService: sharedInfoSvc,
 	})
-
-	options.TerraformVars = map[string]interface{}{
-		"access_tags":               permanentResources["accessTags"],
-		"existing_kms_instance_crn": permanentResources["hpcs_south_crn"],
-		"kms_endpoint_type":         "public",
-		"provider_visibility":       "public",
-		"resource_group_name":       options.Prefix,
-		"admin_pass":                randomPass,
-	}
 
 	output, err := options.RunTestUpgrade()
 	if !options.UpgradeTestSkipped {
